Tests/LibWeb/Text/expected/svg-relayout-sibling-paint-order.txt

hit-before=second
hit-after=second
PASS
